Refactor configuration management by replacing `Settings` with `AgentSettings` and `MCPSettings` across multiple modules. Removed obsolete `config.py` files from `src/agent` and `src/mcp_server`. Updated author information in `pyproject.toml`.
