+++ pyproject.toml
@@ -6,7 +6,7 @@ readme = "README.md"
 requires-python = ">=3.12"
 license = { file = "LICENSE" }
 authors = [
-  { name = "Votre Nom", email = "votre@email.com" },
+  { name = "Andrew ASSEF", email = "assef.andrew@gmail.com" },
 ]
 dependencies = [
     "fastmcp",
