+++ src/core/__init__.py
@@ -1,0 +1,8 @@
+"""
+Module core centralisant la configuration et le logging.
+
+Ce module contient les classes de configuration et les utilitaires
+de logging partagés entre les différents composants de l'application.
+"""
+
+__version__ = "0.1.0"
