+++ src/mcp_server/__init__.py
@@ -5,7 +5,7 @@ Ce module contient tous les composants nécessaires au fonctionnement
 du serveur MCP qui expose l'API data.inclusion.beta.gouv.fr.
 """
 
-from .config import Settings
+from ..core.config import MCPSettings as Settings
 from .server import main
 
 __all__ = ["Settings", "main"]
